Move schema and context under nexus folder
