+++ package.json
@@ -13,7 +13,7 @@
     "prisma:migrate:up": "prisma migrate up --experimental",
     "prisma:db:push": "prisma db push --preview-feature --ignore-migrations --force",
     "prisma:db:seed": "ts-node prisma/seeds/seed",
-    "nexus:reflect": "ts-node --transpile-only api/server/schema",
+    "nexus:reflect": "ts-node --transpile-only api/nexus/schema",
     "generate": "npm run prisma:generate && npm run nexus:reflect",
     "build": "ncc build api/index.ts --quiet --out dist"
   },
